+++ packages/backend/src/socket/DataServer.test.ts
@@ -1,12 +1,12 @@
 import { DataServer } from './DataServer'
-import { getPorts } from '../common/utils'
-import { expect, test } from '@jest/globals'
 
-test.skip('start and stop data server', async () => {
+import { expect, test } from '@jest/globals'
+import { getPorts } from '../common/utils'
+
+test('start and stop data server', async () => {
   const ports = await getPorts()
   const dataServer = new DataServer(ports.dataServer)
-  // TODO:JEST
-  // expect(dataServer.io.engine.opts.cors).toBe(false) // No cors should be set by default
+  expect(dataServer.io.engine.opts.cors).toBe(null) // No cors should be set by default
   await dataServer.listen()
   await dataServer.close()
 })