+++ packages/backend/src/validation/validators.test.ts
@@ -2,7 +2,6 @@ import { ChannelMessage, PublicChannel } from '@quiet/state-manager'
 import { isUser, isMessage, isConversation, isDirectMessage, isChannel } from './validators'
 import { jest, beforeEach, describe, it, expect, afterEach, beforeAll, test } from '@jest/globals'
 
-
 describe('Validators - Users', () => {
   test('publicKey and halfKey are valid', () => {
     const publicKey = '036e4c80bf5defb4ad7798fbe63129aab7c5320a25f19d3e7225edae5a6dd0f079'