+++ packages/backend-bundle/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.19-alpha.13](https://github.com/TryQuiet/quiet/compare/backend-bundle@0.19.19-alpha.12...backend-bundle@0.19.19-alpha.13) (2023-02-08)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
+## [0.19.19-alpha.12](https://github.com/TryQuiet/quiet/compare/backend-bundle@0.19.19-alpha.5...backend-bundle@0.19.19-alpha.12) (2023-02-07)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
+## [0.19.19-alpha.5](https://github.com/TryQuiet/quiet/compare/backend-bundle@0.19.19-alpha.4...backend-bundle@0.19.19-alpha.5) (2023-02-07)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
+## [0.19.19-alpha.4](https://github.com/TryQuiet/quiet/compare/backend-bundle@0.19.19-alpha.3...backend-bundle@0.19.19-alpha.4) (2023-02-06)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
+## [0.19.19-alpha.3](https://github.com/TryQuiet/quiet/compare/backend-bundle@0.19.19-alpha.2...backend-bundle@0.19.19-alpha.3) (2023-02-06)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
+## [0.19.19-alpha.2](https://github.com/TryQuiet/quiet/compare/backend-bundle@0.19.4-alpha.0...backend-bundle@0.19.19-alpha.2) (2023-02-06)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
 ## [0.19.4-alpha.0](https://github.com/TryQuiet/quiet/compare/backend-bundle@0.19.3-alpha.0...backend-bundle@0.19.4-alpha.0) (2023-02-01)
 
 **Note:** Version bump only for package backend-bundle