+++ packages/backend/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.19-alpha.14](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.19-alpha.13...@quiet/backend@0.19.19-alpha.14) (2023-02-08)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+## [0.19.19-alpha.13](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.19-alpha.12...@quiet/backend@0.19.19-alpha.13) (2023-02-07)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+## [0.19.19-alpha.12](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.19-alpha.5...@quiet/backend@0.19.19-alpha.12) (2023-02-07)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+## [0.19.19-alpha.6](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.19-alpha.5...@quiet/backend@0.19.19-alpha.6) (2023-02-07)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+## [0.19.19-alpha.5](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.19-alpha.4...@quiet/backend@0.19.19-alpha.5) (2023-02-06)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+## [0.19.19-alpha.4](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.19-alpha.3...@quiet/backend@0.19.19-alpha.4) (2023-02-06)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+## [0.19.19-alpha.3](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.19-alpha.2...@quiet/backend@0.19.19-alpha.3) (2023-02-06)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+## [0.19.19-alpha.2](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.16-alpha.0...@quiet/backend@0.19.19-alpha.2) (2023-02-06)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
 ## [0.19.16-alpha.0](https://github.com/TryQuiet/backend/compare/@quiet/backend@0.19.15-alpha.0...@quiet/backend@0.19.16-alpha.0) (2023-02-02)
 
 **Note:** Version bump only for package @quiet/backend